+++ src/main.rs
@@ -145,12 +145,7 @@ async fn run(opts: RunOpts) -> Result<(), StatusCodeEnum> {
     loop {
         server_retry_interval.tick().await;
         {
-            match load_data_maybe_empty(
-                i32::from(Regions::Global) as u32,
-                0u64.to_be_bytes().to_vec(),
-            )
-            .await
-            {
+            match load_data_maybe_empty(Regions::Global as u32, 0u64.to_be_bytes().to_vec()).await {
                 Ok(current_block_number_bytes) => {
                     info!("storage service ready, get current height success");
                     if current_block_number_bytes.is_empty() {
@@ -162,7 +157,7 @@ async fn run(opts: RunOpts) -> Result<(), StatusCodeEnum> {
                         current_block_number = u64_decode(current_block_number_bytes);
                         current_block_hash = load_data(
                             storage_client(),
-                            i32::from(Regions::Global) as u32,
+                            Regions::Global as u32,
                             1u64.to_be_bytes().to_vec(),
                         )
                         .await?;
